graphrag-agent-workflow/pyproject.toml
[project]
name = "qa-agent"
version = "0.1.0"
description = "GraphRAG Q&A Agent over Claude Code documentation"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "temporalio>=1.20.0",
    "pydantic>=2.7.0",
    "pydantic-ai>=1.61.0",
    "pydantic-settings>=2.4.0",
    "neo4j>=5.20.0",
    "openai>=1.40.0",
    "cohere>=5.10.0",
    "Flask>=3.0.0",
    "logfire>=4.24.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
]
eval = [
    "ragas>=0.4.3,<0.5",
    "pandas>=2.2,<4",
    "tabulate>=0.9,<1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: integration tests requiring local Neo4j + Ollama (deselect with '-m \"not integration\"')",
]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
